requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.18.0
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
asgiref==3.11.0
asyncer==0.0.8
attrs==25.4.0
backoff==2.2.1
blis==1.3.3
cachetools==6.2.4
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
colorama==0.4.6
colorlog==6.10.1
confection==0.1.5
cryptography==44.0.3
cuid==0.4
cymem==2.0.13
Deprecated==1.3.1
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.17.0
dspy==3.1.0
emoji==2.15.0
fastuuid==0.14.0
filelock==3.20.3
frozenlist==1.8.0
fsspec==2026.1.0
gepa==0.0.22
gitdb==4.0.12
googleapis-common-protos==1.72.0
greenlet==3.3.0
groq==0.37.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
Jinja2==3.1.6
jiter==0.12.0
json_repair==0.44.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==1.2.3
langchain-core==1.2.7
langchain-groq==1.1.1
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.2
langsmith==0.6.2
litellm==1.80.15
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.7.0
murmurhash==1.0.15
netra-sdk==0.1.60
networkx==3.6.1
numpy==2.4.1
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
openai==2.15.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-aio-pika==0.60b1
opentelemetry-instrumentation-aiohttp-client==0.60b1
opentelemetry-instrumentation-aiokafka==0.60b1
opentelemetry-instrumentation-aiopg==0.60b1
opentelemetry-instrumentation-alephalpha==0.47.5
opentelemetry-instrumentation-anthropic==0.47.5
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-asyncclick==0.60b1
opentelemetry-instrumentation-asyncio==0.60b1
opentelemetry-instrumentation-asyncpg==0.60b1
opentelemetry-instrumentation-aws-lambda==0.60b1
opentelemetry-instrumentation-bedrock==0.47.5
opentelemetry-instrumentation-boto3sqs==0.60b1
opentelemetry-instrumentation-botocore==0.60b1
opentelemetry-instrumentation-cassandra==0.60b1
opentelemetry-instrumentation-celery==0.60b1
opentelemetry-instrumentation-chromadb==0.47.5
opentelemetry-instrumentation-click==0.60b1
opentelemetry-instrumentation-cohere==0.47.5
opentelemetry-instrumentation-confluent-kafka==0.60b1
opentelemetry-instrumentation-crewai==0.47.5
opentelemetry-instrumentation-dbapi==0.60b1
opentelemetry-instrumentation-django==0.60b1
opentelemetry-instrumentation-elasticsearch==0.60b1
opentelemetry-instrumentation-falcon==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-flask==0.60b1
opentelemetry-instrumentation-google-generativeai==0.47.5
opentelemetry-instrumentation-groq==0.47.5
opentelemetry-instrumentation-grpc==0.60b1
opentelemetry-instrumentation-haystack==0.47.5
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-instrumentation-jinja2==0.60b1
opentelemetry-instrumentation-kafka-python==0.60b1
opentelemetry-instrumentation-lancedb==0.47.5
opentelemetry-instrumentation-langchain==0.47.5
opentelemetry-instrumentation-llamaindex==0.47.5
opentelemetry-instrumentation-logging==0.60b1
opentelemetry-instrumentation-marqo==0.47.5
opentelemetry-instrumentation-mcp==0.47.5
opentelemetry-instrumentation-milvus==0.47.5
opentelemetry-instrumentation-mistralai==0.47.5
opentelemetry-instrumentation-mysql==0.60b1
opentelemetry-instrumentation-mysqlclient==0.60b1
opentelemetry-instrumentation-ollama==0.47.5
opentelemetry-instrumentation-openai==0.47.5
opentelemetry-instrumentation-openai-agents==0.47.5
opentelemetry-instrumentation-pika==0.60b1
opentelemetry-instrumentation-pinecone==0.47.5
opentelemetry-instrumentation-psycopg==0.60b1
opentelemetry-instrumentation-psycopg2==0.60b1
opentelemetry-instrumentation-pymemcache==0.60b1
opentelemetry-instrumentation-pymongo==0.60b1
opentelemetry-instrumentation-pymssql==0.60b1
opentelemetry-instrumentation-pymysql==0.60b1
opentelemetry-instrumentation-qdrant==0.47.5
opentelemetry-instrumentation-redis==0.60b1
opentelemetry-instrumentation-remoulade==0.60b1
opentelemetry-instrumentation-replicate==0.47.5
opentelemetry-instrumentation-requests==0.60b1
opentelemetry-instrumentation-sagemaker==0.47.5
opentelemetry-instrumentation-sqlalchemy==0.60b1
opentelemetry-instrumentation-sqlite3==0.60b1
opentelemetry-instrumentation-starlette==0.60b1
opentelemetry-instrumentation-system-metrics==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-instrumentation-together==0.47.5
opentelemetry-instrumentation-tornado==0.60b1
opentelemetry-instrumentation-tortoiseorm==0.60b1
opentelemetry-instrumentation-transformers==0.47.5
opentelemetry-instrumentation-urllib==0.60b1
opentelemetry-instrumentation-urllib3==0.60b1
opentelemetry-instrumentation-vertexai==0.47.5
opentelemetry-instrumentation-watsonx==0.47.5
opentelemetry-instrumentation-weaviate==0.47.5
opentelemetry-instrumentation-writer==0.47.5
opentelemetry-instrumentation-wsgi==0.60b1
opentelemetry-propagator-aws-xray==1.0.2
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.60b1
optuna==4.6.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
phonenumbers==8.13.55
posthog==3.25.0
preshed==3.0.12
presidio_analyzer==2.2.358
presidio_anonymizer==2.2.358
propcache==0.4.1
protobuf==6.33.4
psutil==7.2.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-file==3.0.1
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
safetensors==0.7.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
smmap==5.0.2
sniffio==1.3.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.45
srsly==2.5.2
stanza==1.11.0
sympy==1.14.0
tenacity==9.1.2
thinc==8.3.10
tiktoken==0.12.0
tldextract==5.3.1
tokenizers==0.21.4
torch==2.9.1
tqdm==4.67.1
traceloop-sdk==0.47.5
transformers==4.51.3
triton==3.5.1
typer-slim==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
wasabi==1.1.3
weasel==0.4.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
